% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkema.R
\docType{data}
\name{arkema}
\alias{arkema}
\alias{arkema.model}
\title{Effects of giant kelp on benthic community structure}
\format{A data frame with 69 observations of 8 variables}
\source{
Article:

\url{http://onlinelibrary.wiley.com/doi/10.1890/08-1213.1/abstract}
}
\usage{
data(arkema)
arkema
arkema.model
}
\description{
Raw data and specification of the SEM model in Fig. 5 in Arkema et al. 2009 (Ecology)
}
\examples{
\dontrun{
data(arkema)
summary(sem(arkema.model, arkema), standardized = TRUE)
}
}
\references{
Arkema et al. 2009. Direct and indirect effects of giant kelp determine
benthic community structure and dynamics. Ecology 11: 3126-3137.
}

