% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{semeco}
\alias{semeco}
\alias{semeco-package}
\title{Package semeco}
\description{
Examples of structural equation modelling usage in Ecology
}
\details{
This package contains data and specified structural models published in
ecological journals. Models follow the lavaan syntax
}

